[Vulkan]Fix compile error

+++ include/bksge/core/render/vulkan/detail/inl/resource_pool_inl.hpp
@@ -16,13 +16,14 @@
 #include <bksge/core/render/vulkan/detail/shader.hpp>
 #include <bksge/core/render/vulkan/detail/geometry.hpp>
 #include <bksge/core/render/vulkan/detail/graphics_pipeline.hpp>
-#include <bksge/core/render/vulkan/detail/vulkan_h.hpp>
+#include <bksge/core/render/vulkan/detail/vulkan.hpp>
 #include <bksge/core/render/shader.hpp>
 #include <bksge/core/render/geometry.hpp>
 #include <bksge/core/render/render_state.hpp>
 #include <bksge/fnd/functional/hash_combine.hpp>
 #include <bksge/fnd/utility/forward.hpp>
 #include <memory>	// make_shared
+#include <cstdint>
 
 namespace bksge
 {
@@ -96,7 +97,7 @@ ResourcePool::GetGraphicsPipeline(
 	vulkan::UniformBuffer* uniform_buffer)
 {
 	auto const id = bksge::hash_combine(
-		num_samples,
+		static_cast<std::uint32_t>(num_samples),
 		geometry.id(),
 		shader.id(),
 		render_state);
